Remove legacy storage repository aggregate test
