+++ tests/worldservice/storage/test_validation_cache_repository.py
@@ -2,67 +2,9 @@ from __future__ import annotations
 
 import pytest
 
+from qmtl.worldservice.storage.audit import AuditLogRepository
 from qmtl.worldservice.storage.models import ValidationCacheEntry
-from qmtl.worldservice.storage.repositories import (
-    AuditLogRepository,
-    EdgeOverrideRepository,
-    ValidationCacheRepository,
-    WorldRepository,
-    _REASON_UNSET,
-)
-
-
-def test_world_repository_persists_records_and_audits() -> None:
-    audit = AuditLogRepository()
-    repo = WorldRepository(audit)
-
-    record = repo.create({"id": "world-1", "name": "First"})
-    assert record.id == "world-1"
-    assert repo.get("world-1") == {"id": "world-1", "name": "First"}
-
-    repo.update("world-1", {"description": "primary"})
-    assert repo.get("world-1") == {
-        "id": "world-1",
-        "name": "First",
-        "description": "primary",
-    }
-
-    entries = audit.list_entries("world-1")
-    assert [event["event"] for event in entries] == [
-        "world_created",
-        "world_updated",
-    ]
-
-
-def test_edge_override_repository_reason_round_trip() -> None:
-    audit = AuditLogRepository()
-    repo = EdgeOverrideRepository(audit)
-
-    created = repo.upsert(
-        "world-1",
-        "domain:backtest",
-        "domain:live",
-        active=False,
-        reason="auto:block",
-    )
-    assert created.reason == "auto:block"
-
-    preserved = repo.upsert(
-        "world-1",
-        "domain:backtest",
-        "domain:live",
-        active=True,
-        reason=_REASON_UNSET,
-    )
-    assert preserved.reason == "auto:block"
-
-    repo.delete("world-1", "domain:backtest", "domain:live")
-    events = audit.list_entries("world-1")
-    assert [event["event"] for event in events] == [
-        "edge_override_upserted",
-        "edge_override_upserted",
-        "edge_override_deleted",
-    ]
+from qmtl.worldservice.storage.validation_cache import ValidationCacheRepository
 
 
 @pytest.mark.parametrize("domain", ["BACKTEST", "Backtest", "backtest"])
